encoder-decoder-model/requirements.txt
absl-py==0.8.0
asn1crypto==0.24.0
astor==0.8.0
certifi==2020.4.5.1
chardet==3.0.4
cryptography==2.1.4
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
Markdown==3.1.1
matplotlib
numpy==1.17.2
opt-einsum==3.0.1
protobuf==3.9.2
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
python-apt==1.6.4
python-dateutil==2.8.1
pyxdg==0.25
requests==2.23.0
scipy==1.4.1
SecretStorage==2.3.1
six==1.11.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0rc1
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==0.16.0
wrapt==1.11.2
sklearn
